fix(epytodo): fixing todo issues

+++ src/routes/todos/todos.query.js
@@ -1,21 +1,22 @@
 const db = require('../../config/db');
 
 exports.getAllTodos = (req, res) => {
-    const userId = req.user.id;
-    db.query('SELECT * FROM todo WHERE user_id = ?', [userId], (err, results) => {
+    db.query('SELECT * FROM todo', (err, results) => {
     if (err)
-        return res.status(500).json({ error: 'DB error' });
+        return res.status(500).json({ msg: 'DB error' });
     return res.status(200).json(results);
   });
 };
 
 exports.getTodoById = (req, res) => {
   const { id } = req.params;
+  if (!id || isNaN(id))
+    return res.status(400).json({ msg: 'Bad parameters' });
   db.query('SELECT * FROM todo WHERE id = ?', [id], (err, results) => {
     if (err)
-      return res.status(500).json({ error: 'DB error' });
+      return res.status(500).json({ msg: 'DB error' });
     if (results.length === 0)
-      return res.status(404).json({ error: 'Todo not found' });
+      return res.status(404).json({ msg: 'Todo not found' });
     return res.status(200).json(results[0]);
   });
 };
@@ -23,12 +24,12 @@ exports.getTodoById = (req, res) => {
 exports.createTodo = (req, res) => {
     const { title, description, due_time, user_id, status } = req.body;
 
-    if (!title || !description || !due_time || !user_id || !status) {
+    if (!title || !description || !due_time || !user_id || !status || status === undefined) {
         return res.status(400).json({ msg: 'Bad parameters' });
     }
     db.query('SELECT * FROM user WHERE id = ?', [user_id], (err, results) => {
         if (err) {
-            return res.status(500).json({ error: 'DB error' });
+            return res.status(500).json({ msg: 'DB error' });
         }
         if (results.length === 0) {
             return res.status(404).json({ msg: 'User not found' });
@@ -38,7 +39,7 @@ exports.createTodo = (req, res) => {
             [title, description, due_time, user_id, status || 'not started'],
             (err, result) => {
                 if (err) {
-                    return res.status(500).json({ error: 'DB error' });
+                    return res.status(500).json({ msg: 'DB error' });
                 }
                 return res.status(201).json({ id: result.insertId, title, description, due_time, user_id, status });
             }
@@ -49,7 +50,7 @@ exports.createTodo = (req, res) => {
 exports.updateTodo = (req, res) => {
     const { id } = req.params;
     const { title, description, due_time, user_id, status } = req.body;
-    if (!title || !description || !due_time || !user_id || !status) {
+    if (!id || !title || !description || !due_time || !user_id || !status || status === undefined) {
         return res.status(400).json({ msg: 'Bad parameters' });
     }
     db.query(
@@ -57,7 +58,7 @@ exports.updateTodo = (req, res) => {
     [title, description, due_time, user_id, status, id],
     (err, results) => {
         if (err) 
-          return res.status(500).json({ error: 'DB error' });
+          return res.status(500).json({ msg: 'DB error' });
         if (results.length === 0)
           return res.status(404).json({ msg: 'Todo not found' });
         return res.status(200).json({ id: id, title, description, due_time, user_id, status });
@@ -69,9 +70,17 @@ exports.deleteTodo = (req, res) => {
     const { id } = req.params;
     if (!id)
         return res.status(400).json({ msg: 'Bad parameters' });
-    db.query('DELETE FROM todo WHERE id = ?', [id], (err) => {
-    if (err)
-        return res.status(500).json({ error: 'DB error' });
-    return res.status(200).json({ msg: `Successfully deleted record number : ${id}` });
-  });
+    db.query('SELECT * FROM todo WHERE id = ?', [id], (err, results) => {
+      if (err) {
+        return res.status(500).json({ msg: 'DB error' });
+      }
+      if (results.length === 0) {
+        return res.status(404).json({ msg: 'Todo not found' });
+      }
+      db.query('DELETE FROM todo WHERE id = ?', [id], (err) => {
+      if (err)
+        return res.status(500).json({ msg: 'DB error' });
+      return res.status(200).json({ msg: `Successfully deleted record number : ${id}` });
+      });
+    });
 };
